Extract fixtures used across multiple test files.
